requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
rank-bm25
peft
